refactor platform specific code out of device.dart
remove device type specific checks
