+++ packages/flutter_tools/lib/src/commands/trace.dart
@@ -4,9 +4,9 @@
 
 import 'dart:async';
 
+import '../android/device_android.dart';
 import '../application_package.dart';
 import '../base/logging.dart';
-import '../device.dart';
 import '../runner/flutter_command.dart';
 
 class TraceCommand extends FlutterCommand {
