+++ packages/flutter_tools/test/android_device_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter_tools/src/device.dart';
+import 'package:flutter_tools/src/android/device_android.dart';
 import 'package:test/test.dart';
 
 main() => defineTests();
